requirements.txt
torch==2.2.1
accelerate==0.30.0
albumentations==1.4.6
flash_attn==2.6.1
h5py==3.11.0
huggingface_hub==0.23.3
hydra-core==1.3.2
matplotlib==3.8.4
numpy==2.1.2
omegaconf==2.3.0
opencv_python==4.9.0.80
opencv_python_headless==4.9.0.80
pandas==2.2.3
Pillow==11.0.0
safetensors==0.4.5
scikit_learn==1.4.2
scipy==1.14.1
scikit_image==0.23.2
xformers==0.0.25
rocket @ git+https://github.com/dsenushkin/rocket.git@b965da378b3d4f9d41c60dc0548cfb3657d888eb